requirements.txt
streamlit
langchain-google-genai
langgraph
requests
python-dotenv